+++ shuttlelog/src/main/java/edu/bsu/shuttlelog/service/LogService.java
@@ -1,5 +1,6 @@
 package edu.bsu.shuttlelog.service;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import edu.bsu.shuttlelog.entity.Log;
@@ -7,7 +8,7 @@ import edu.bsu.shuttlelog.entity.Log;
 public interface LogService {
 	List<Log> list();
 
-	List<Long> save(List<Log> log);
+	List<BigInteger> save(List<Log> log);
 
 	Log getByID(long id);
